# 8273100: Improve AbstractStringBuilder.append(String) when using CompactStrings