@cl4es 2021-08-30
Simplified further after realizing `putStringAt(int, String)` can use `String.getBytes(String, int, byte)`, improving performance for both modes:

```
# +CompactStrings
Benchmark                    Mode  Cnt    Score    Error  Units
StringBuilders.appendLoop16  avgt   15  681.778 ± 14.978  ns/op
StringBuilders.appendLoop8   avgt   15  358.008 ±  9.230  ns/op

# -CompactStrings
Benchmark                    Mode  Cnt    Score    Error  Units
StringBuilders.appendLoop16  avgt   15  664.591 ± 10.505  ns/op
StringBuilders.appendLoop8   avgt   15  351.175 ± 13.282  ns/op

```