@stsypanov 2021-08-30
Hi, just curious how have you found out that the code should be extracted into a separate methods? Profiler?