@cl4es 2021-08-30
> Hi, just curious how have you found out that the code should be extracted into a separate methods? Profiler?

I saw that `String::length` calls appeared more than once with async-profiler, then did some experiments to see if manually inlining so that only one call was needed helped (and it did).